fix a theoretical CWE-190 integer overflow as out.length is a long while the for loop is indexed by an int https://docs.oracle.com/javase/8/docs/api/index.html?java/io/RandomAccessFile.html

+++ src/main/java/org/owasp/esapi/waf/internal/InterceptingServletOutputStream.java
@@ -117,7 +117,7 @@ public class InterceptingServletOutputStream extends ServletOutputStream {
 				
 				byte[] buff = new byte[FLUSH_BLOCK_SIZE];
 				
-				for(int i=0;i<out.length();) {
+				for(long i=0;i<out.length();) {
 					
 					long currentPos = out.getFilePointer();
 					long totalSize = out.length();
